added DumpTable() and semiworking FilterRoutes()
I implemented the NodeStore_dumpTable() function under the shorter name
of DumpTable(). It provides the same function.

I also added a FilterRoutes() function which is able to filter routes
based on target IP address. Soon, it will also be able to remove routes
based on distance (in number of hops) to that node.